+++ packages/core/lib/match.js
@@ -1,8 +1,13 @@
+const _ = require('lodash')
 const globToRegex = require('glob-to-regexp')
 const { assertConfigSetters, invokeConfigSetters } = require('./configSetters')
 const { invokePreHooks, invokePostHooks } = require('./hooks')
 
 module.exports = match
+
+const regexify = glob => globToRegex(glob, { extended: true })
+const stripArrayConditionally = array => array.length === 1 ? array[0] : array
+const toArray = thing => Array.isArray(thing) ? thing : [ thing ]
 
 /**
  * State on which files to apply the loader blocks passed in this call. Works
@@ -23,13 +28,17 @@ function match (test, options, configSetters) {
 
   assertConfigSetters(configSetters)
 
-  const match = { test: createFileTypeMatcher(test) }
+  const { inclusions, exclusions } = splitPatterns(toArray(test))
+  const match = Object.assign({}, options, {
+    // The `.test` is usually just one pattern, so stripping the array feels more natural
+    test: stripArrayConditionally(normalizeMatchers(inclusions))
+  })
 
-  if (options.exclude) {
-    match.exclude = options.exclude
-  }
-  if (options.include) {
-    match.include = options.include
+  if (exclusions.length > 0) {
+    match.exclude = _.concat(
+      match.exclude ? toArray(match.exclude) : [],
+      normalizeMatchers(exclusions)
+    )
   }
 
   const groupBlock = context => config => invokeConfigSetters(configSetters, deriveContextWithMatch(context, match), config)
@@ -40,15 +49,23 @@ function match (test, options, configSetters) {
   })
 }
 
-const regexify = glob => globToRegex(glob, { extended: true })
+function normalizeMatchers (fileMatchTests) {
+  return fileMatchTests.map(test => {
+    if (typeof test === 'string') {
+      return regexify(test)
+    } else {
+      return test
+    }
+  })
+}
 
-function createFileTypeMatcher (test) {
-  if (typeof test === 'string') {
-    return regexify(test)
-  } else if (Array.isArray(test) && test.every(item => typeof item === 'string')) {
-    return test.map(item => regexify(item))
-  } else {
-    return test
+function splitPatterns (patterns) {
+  const isNegation = pattern => typeof pattern === 'string' && pattern.startsWith('!')
+  const stripLeadingExclam = pattern => pattern.startsWith('!') ? pattern.substr(1) : pattern
+
+  return {
+    inclusions: patterns.filter(pattern => !isNegation(pattern)),
+    exclusions: patterns.filter(pattern => isNegation(pattern)).map(stripLeadingExclam)
   }
 }